Add confirm dialog when deleting a user from admin

+++ client/app/admin/admin.component.ts
@@ -31,11 +31,13 @@ export class AdminComponent implements OnInit {
   }
 
   deleteUser(user) {
-    this.userService.deleteUser(user).subscribe(
-      data => this.toast.setMessage('user deleted successfully.', 'success'),
-      error => console.log(error),
-      () => this.getUsers()
-    );
+    if (window.confirm('Are you sure you want to delete ' + user.username + '?')) {
+      this.userService.deleteUser(user).subscribe(
+        data => this.toast.setMessage('user deleted successfully.', 'success'),
+        error => console.log(error),
+        () => this.getUsers()
+      );
+    }
   }
 
 }
